fix(data-warehouse): keep the google-ads SDK off the source-registration path
Registering data-import sources imports every source module so each self-registers with
SourceRegistry. The google-ads source pulled the google-ads SDK at module scope, so registering
ANY source (e.g. a Postgres sync calling get_source) imported google-ads too. That SDK has a
circular module graph (services <-> common <-> resources) only safe to import single-threaded;
under concurrent worker-thread imports it deadlocks or duplicate-registers its protobuf
descriptors, surfacing as 'Couldn't build proto file into descriptor pool: duplicate symbol
google.ads.googleads.v23...'. Once #60635 moved source loading to runtime, unrelated syncs began
tripping it.

Split the lightweight config dataclasses into configs.py (no SDK) and defer the SDK-backed
functions in source.py into the methods that run them. Registering a source no longer imports
google-ads; the SDK loads only when a google-ads sync actually runs, through the single google_ads
module whose per-module import lock serializes concurrent first-imports (no deadlock). A guard test
asserts source registration stays SDK-free.

Validated against a local Postgres + Temporal stack: get_source(Postgres) no longer imports
google.ads, and concurrent SDK imports through the entry module complete without the circular-import
deadlock the pre-split direct imports hit.
