+++ posthog/temporal/data_imports/sources/google_ads/google_ads.py
@@ -1,6 +1,5 @@
 import typing
 import datetime as dt
-import dataclasses
 import collections.abc
 
 from django.conf import settings
@@ -20,61 +19,17 @@ from posthog.models.integration import Integration
 from posthog.temporal.data_imports.naming_convention import NamingConvention
 from posthog.temporal.data_imports.pipelines.helpers import incremental_type_to_initial_value
 from posthog.temporal.data_imports.pipelines.pipeline.typings import SourceResponse
-from posthog.temporal.data_imports.sources.common import config
 from posthog.temporal.data_imports.sources.common.resumable import ResumableSourceManager
 from posthog.temporal.data_imports.sources.common.sql import Column, Table
 from posthog.temporal.data_imports.sources.generated_configs import GoogleAdsSourceConfig
+from posthog.temporal.data_imports.sources.google_ads.configs import (
+    GoogleAdsResumeConfig,
+    GoogleAdsSourceConfigUnion,
+    clean_customer_id,
+)
 from posthog.temporal.data_imports.sources.google_ads.schemas import FIELD_ALIASES, RESOURCE_SCHEMAS
 
 from products.data_warehouse.backend.types import IncrementalFieldType
-
-
-@dataclasses.dataclass
-class GoogleAdsResumeConfig:
-    """Resumable state for the Google Ads source.
-
-    `page_token` is the opaque continuation token returned by
-    `GoogleAdsService.search` for the next page to fetch.
-    """
-
-    page_token: str
-
-
-def clean_customer_id(s: str | None) -> str | None:
-    """Clean customer IDs from Google Ads.
-
-    Customer IDs can contain dashes, but we need the ID without them.
-    """
-    if not s:
-        return s
-
-    return s.strip().replace("-", "")
-
-
-@config.config
-class GoogleAdsServiceAccountSourceConfig(config.Config):
-    """Google Ads source config using service account for authentication.
-
-    Old config for when we were using a service account instead of oauth.
-    ~100 sources still use this method for auth. We recommend using
-    `GoogleAdsSourceConfig` instead"""
-
-    customer_id: str = config.value(converter=clean_customer_id)
-
-    private_key: str = config.value(
-        default_factory=config.default_from_settings("GOOGLE_ADS_SERVICE_ACCOUNT_PRIVATE_KEY")
-    )
-    private_key_id: str = config.value(
-        default_factory=config.default_from_settings("GOOGLE_ADS_SERVICE_ACCOUNT_PRIVATE_KEY_ID")
-    )
-    client_email: str = config.value(
-        default_factory=config.default_from_settings("GOOGLE_ADS_SERVICE_ACCOUNT_CLIENT_EMAIL")
-    )
-    token_uri: str = config.value(default_factory=config.default_from_settings("GOOGLE_ADS_SERVICE_ACCOUNT_TOKEN_URI"))
-    developer_token: str = config.value(default_factory=config.default_from_settings("GOOGLE_ADS_DEVELOPER_TOKEN"))
-
-
-GoogleAdsSourceConfigUnion = GoogleAdsServiceAccountSourceConfig | GoogleAdsSourceConfig
 
 
 def google_ads_client(config: GoogleAdsSourceConfigUnion, team_id: int) -> GoogleAdsClient:
